Responded to code review
